chore: Minor refactors
