+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/FieldInfo.java
@@ -1,0 +1,3 @@
+package com.appsmith.server.dtos;
+
+public record FieldInfo(String fullPath, Class<?> type) {}
